TabSplitter/requirements.txt
attrs==23.1.0
PyGuitarPro==0.9.3
